Cleanup and prep
Moved helpers to appropriate location.

Created WPF windows.

Going to start designing.
