+++ MusicDLP.csproj
@@ -55,7 +55,6 @@
   </ItemGroup>
   <ItemGroup>
     <Folder Include="Model\" />
-    <Folder Include="ViewModel\Helpers\" />
     <Folder Include="ViewModel\Commands\" />
   </ItemGroup>
   <ItemGroup>
