+++ MusicDLP.csproj.user
@@ -47,4 +47,12 @@
       <SubType>Designer</SubType>
     </None>
   </ItemGroup>
+  <ItemGroup>
+    <Page Update="View\OptionsForm.xaml">
+      <SubType>Designer</SubType>
+    </Page>
+    <Page Update="View\CustomDialogBox.xaml">
+      <SubType>Designer</SubType>
+    </Page>
+  </ItemGroup>
 </Project>
