+++ View/CustomDialogBox.xaml.cs
@@ -1,0 +1,9 @@
+﻿using System.Windows;
+
+namespace MusicDLP.View;
+
+public partial class CustomDialogBox : Window
+{
+    public CustomDialogBox() => InitializeComponent();
+
+}
