+++ View/OptionsForm.xaml
@@ -1,0 +1,13 @@
+﻿<Window x:Class="MusicDLP.View.OptionsForm"
+              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+              xmlns:i="http://metro.mahapps.com/winfx/xaml/iconpacks"
+              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+              xmlns:local="clr-namespace:MusicDLP.View"
+              mc:Ignorable="d"
+              Style="{StaticResource CustomWindowStyle}">
+    <DockPanel>
+
+    </DockPanel>
+</Window>
