+++ View/OptionsForm.xaml.cs
@@ -1,0 +1,9 @@
+﻿using System.Windows;
+
+namespace MusicDLP.View;
+
+public partial class OptionsForm : Window
+{
+    public OptionsForm() => InitializeComponent();
+
+}
