+++ ViewModel/Helpers/URLConstants.cs
@@ -1,0 +1,8 @@
+﻿namespace MusicDLP.ViewModel.Helpers
+{
+    public class URLConstants
+    {
+        public static string ytdlpDownloadUrl = "https://github.com/yt-dlp/yt-dlp/releases/download/2023.01.06/yt-dlp.exe";
+        public static string FFMPEGDownloadUrl = "https://github.com/BtbN/FFmpeg-Builds/releases/download/latest/ffmpeg-master-latest-win64-lgpl.zip";
+    }
+}
